Added a temporary desktop hack as the graph is just unviewable given the crazy scale on desktop. Also updated all the labels to use the new matrix format, and moved my radar graph to its own component.
