+++ src/index.css
@@ -79,6 +79,14 @@ textarea:hover {
   outline: 2px solid #535bf2;
 }
 
+/* This is a quick temporary hack to make the desktop view easier on the eyes */
+@media screen and (min-width: 800px) {
+  .contentContainer {
+    width: 50%;
+    margin: 0 auto;
+  }
+}
+
 @media (prefers-color-scheme: light) {
   :root {
     color: #213547;
